Ship MemoryCache
Thanks @leplatrem for the suggestion.
